+++ src/routes/moderator.py
@@ -38,8 +38,13 @@ async def delete_photo(
     Returns:
         dict: The deleted photo details.
     """
-    photo = await repository_photo.delete_photo(current_user.id, photo_id, db)
-
+    if current_user.role_id == 2 or current_user.role_id == 1:
+        photo = await repository_photo.delete_photo(current_user.id, photo_id, db)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=messages.ACCESS_FORBIDDEN,
+        )
     return photo
 
 
@@ -65,26 +70,31 @@ async def update_comment(
     Returns:
         dict: The updated comment details.
     """
-
-    photo = db.query(Photo).filter(Photo.id == photo_id).first()
-    if not photo:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail=messages.PHOTO_NOT_FOUND
-        )
-    else:
-        comment = (
-            db.query(Comment)
-            .filter(
-                Comment.id == comment_id,
-                Comment.photo_id == photo_id,
-                Comment.user_id == current_user.id,
-            )
-            .first()
-        )
-        if not comment:
+    if current_user.role_id == 2 or current_user.role_id == 1:
+        photo = db.query(Photo).filter(Photo.id == photo_id).first()
+        if not photo:
             raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail=messages.COMMENT_NOT_FOUND
+                status_code=status.HTTP_404_NOT_FOUND, detail=messages.PHOTO_NOT_FOUND
             )
+        else:
+            comment = (
+                db.query(Comment)
+                .filter(
+                    Comment.id == comment_id,
+                    Comment.photo_id == photo_id,
+                    Comment.user_id == current_user.id,
+                )
+                .first()
+            )
+            if not comment:
+                raise HTTPException(
+                    status_code=status.HTTP_404_NOT_FOUND, detail=messages.COMMENT_NOT_FOUND
+                )
 
-        comment = update_comment_rep(db, comment_id, updated_comment)
-        return comment
+            comment = update_comment_rep(db, comment_id, updated_comment)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=messages.ACCESS_FORBIDDEN,
+        )
+    return comment
